Refatorado calculo do ingresso.

+++ src/br/edu/SistemaDeEventos/modelo/Publico.java
@@ -2,6 +2,6 @@ package br.edu.SistemaDeEventos.modelo;
 
 public interface Publico {
 
-	public Integer getPercentualDesconto();
+	public Double getPercentualDesconto();
 	
 }
